chore: rename some remaining cwf and web styles mentions
