+++ .changeset/violet-snakes-design.md
@@ -3,5 +3,5 @@
 '@swisspost/design-system-styles': patch
 ---
 
-Created a new package offering Migration Schematics for the Web Styles
+Created a new package offering Migration Schematics for the Design System Styles
 Removed custom input styles that are no longer used
