+++ packages/demo/src/app/app.module.ts
@@ -24,7 +24,7 @@ import { IntranetComponentsModule } from "./intranet-layout/intranet-components.
 import { PostSampleComponentsModule } from "./post-sample/post-sample-components.module";
 import { SwissPostIntranetHeaderModule } from "@swisspost/design-system-intranet-header";
 import { HomeComponent } from "./home/home.component";
-import { PostCommonModule } from "./common/cwf-common.module";
+import { PostCommonModule } from "./common/post-common.module";
 
 registerLocaleData(localeDe);
 registerLocaleData(localeFr);
