+++ packages/demo/src/app/home/home.component.html
@@ -112,10 +112,10 @@
                 <p>Check if new versions are compatible with your project.</p>
               </li>
               <li>
-                <h3 class="font-curve-regular bold">Prevent known issue with Angular 12 and CWF</h3>
+                <h3 class="font-curve-regular bold">Prevent known issue with Angular 12 and the Design System</h3>
                 <p>
                   There is a known issue with the url() CSS property and the new Angular build
-                  optimizer. CWF is also affected so you need to disable the "inlineCritical"
+                  optimizer. The Design System is also affected so you need to disable the "inlineCritical"
                   optimization for the time being. Head over to your
                   <code>angular.json</code>
                   and set
@@ -223,7 +223,7 @@
       <div class="row pt-huge-r">
         <div class="topic-teaser-content col-12 col-lg-8">
           <h2 class="topic-teaser-title font-curve-large mb-large">
-            <span class="bold">CWF Resources</span>
+            <span class="bold">Design System Resources</span>
             <span class="light">For a unified, accessible user experience</span>
           </h2>
           <ul class="link-list font-curve-regular mb-large">
